adding exceptions, validations

+++ HR.LeaveManagement.Application/Features/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,7 @@
 using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
+using HR.LeaveManagement.Domain;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.Commands.DeleteLeaveType;
@@ -13,10 +15,17 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
     
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
+        // retrieve domain entity object
         var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
         
+        // verify that record exists
+        if (leaveTypeToDelete == null)
+            throw new NotFoundException(nameof(LeaveType), request.Id);
+        
+        // remove from database
         await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
         
+        // return record id
         return Unit.Value; 
     }
 }
